refactor(portal): inline the dormant account check per caller
