+++ elixir/lib/portal/workers/outdated_gateways.ex
@@ -12,7 +12,6 @@ defmodule Portal.Workers.OutdatedGateways do
   require Logger
 
   alias __MODULE__.Database
-  alias Portal.Accounts.Activity
   alias Portal.{Device, Mailer}
 
   @impl Oban.Worker
@@ -25,7 +24,6 @@ defmodule Portal.Workers.OutdatedGateways do
     latest_version = Portal.ComponentVersions.gateway_version()
 
     Database.all_accounts_pending_notification!()
-    |> reject_dormant_accounts()
     |> Enum.each(fn account ->
       incompatible_client_count = Database.count_incompatible_for(account, latest_version)
 
@@ -33,12 +31,6 @@ defmodule Portal.Workers.OutdatedGateways do
       |> Enum.filter(&Device.gateway_outdated?/1)
       |> send_notifications(account, incompatible_client_count)
     end)
-  end
-
-  defp reject_dormant_accounts(accounts) do
-    active_ids = Activity.active_account_ids(Enum.map(accounts, & &1.id))
-
-    Enum.filter(accounts, &MapSet.member?(active_ids, &1.id))
   end
 
   defp all_online_gateways_for_account(account) do
@@ -129,8 +121,11 @@ defmodule Portal.Workers.OutdatedGateways do
     alias Portal.Safe
     alias Portal.Device
 
+    # Accounts with no session logs are dormant and are left out entirely, so
+    # their last_notified stays unset and they are notified once they return.
     def all_accounts_pending_notification! do
       from(a in Portal.Account,
+        as: :accounts,
         where: fragment("?->'notifications'->'outdated_gateway'->>'enabled' = 'true'", a.config),
         where:
           fragment(
@@ -140,7 +135,14 @@ defmodule Portal.Workers.OutdatedGateways do
             fragment(
               "(?->'notifications'->'outdated_gateway'->>'last_notified')::timestamp < timezone('UTC', NOW()) - interval '24 hours'",
               a.config
+            ),
+        where:
+          exists(
+            from(sl in Portal.SessionLog,
+              where: sl.account_id == parent_as(:accounts).id,
+              select: 1
             )
+          )
       )
       |> Safe.unscoped()
       |> Safe.all()
